Move b.err check out of finish
This was added in da68abe58f05f9a0b82eafc0ecbdb4f7323cc242 to allow
Builder.Bytes to be inlined in go1.12. This is no longer possible as
5abc5987da676e7685a13fa59d1fdcd24f88b486 added a cast to Builder.Bytes
and Builder.BytesOrPanic that pushes them past the inlining budget.

Removing this makes builder.finish simpler and more consistent with the
rest of the exported and unexported API.